Fixed some DUMB bugs

+++ Bengine/GUI.cpp
@@ -183,7 +183,7 @@ void Bengine::GUI::onSDLEvent(SDL_Event& evnt) {
     CEGUI::utf32 codePoint;
     switch (evnt.type) {
         case SDL_MOUSEMOTION:
-            m_context->injectMousePosition(evnt.motion.x, evnt.motion.y);
+            m_context->injectMousePosition((float)evnt.motion.x, (float)evnt.motion.y);
             break;
         case SDL_KEYDOWN:
             m_context->injectKeyDown(SDLKeyToCEGUIKey(evnt.key.keysym.sym));
